Modal window in an independent component
