--- js-storefront/taloscommercestore/src/app/spartacus/features/tc-saved-list/components/tc-saved-list-pdp-btn-add/tc-saved-list-pdp-btn-add.component.scss
@@ -1,36 +1,0 @@
-.tc-modal__button {
-  padding-top: 20px;
-  padding-inline-end: 20px;
-  padding-bottom: 0px;
-  padding-inline-start: 20px;
-}
-
-label{
-  width: 100%;
-}
-
-.error{
-  display: flex; 
-  align-items: center;
-}
-
-.error-icon{
-  color: red;
-}
-
-.message-error{
-  margin: auto 5px;
-}
-
-.error-container{
-  margin-top: 5px;
-}
-
-.close-icon{
-  color: #b4b4b4;
-}
-
-
-.close-icon:hover{
-  color: black;
-}
