+++ js-storefront/taloscommercestore/src/app/spartacus/features/tc-saved-list/components/tc-saved-list-pdp-btn-add/tc-saved-list-pdp-btn-add.module.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { TcSavedListPdpBtnAddComponent } from './tc-saved-list-pdp-btn-add.component';
-import { CmsConfig, ConfigModule, I18nModule } from '@spartacus/core';
-import { IconModule} from '@spartacus/storefront';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NgSelectModule } from '@ng-select/ng-select';
+import { CmsConfig, ConfigModule } from '@spartacus/core';
+import { TcSavedListModalModule } from 'src/app/spartacus/shared/cms-components/content/tc-saved-list-modal/tc-saved-list-modal.module';
 
 
 
@@ -12,12 +10,7 @@ import { NgSelectModule } from '@ng-select/ng-select';
   declarations: [TcSavedListPdpBtnAddComponent],
   imports: [
     CommonModule,
-    IconModule,
-    ReactiveFormsModule,
-    NgSelectModule, 
-    FormsModule,
-    I18nModule,
-    IconModule,
+    TcSavedListModalModule,
     ConfigModule.forRoot({
         cmsComponents: {
             StockNotificationComponent: {
