+++ js-storefront/taloscommercestore/src/app/spartacus/shared/cms-components/content/tc-saved-list-modal/assets/translations/translations.ts
@@ -1,0 +1,5 @@
+import {TranslationChunksConfig} from '@spartacus/assets';
+
+export const tcSavedListTranslationChunksConfig: TranslationChunksConfig = {
+  tcSavedList: ['savedListForm'],
+};
