+++ js-storefront/taloscommercestore/src/app/spartacus/shared/cms-components/content/tc-saved-list-modal/tc-saved-list-modal.component.scss
@@ -1,0 +1,36 @@
+.tc-modal__button {
+    padding-top: 20px;
+    padding-inline-end: 20px;
+    padding-bottom: 0px;
+    padding-inline-start: 20px;
+  }
+  
+  label{
+    width: 100%;
+  }
+  
+  .error{
+    display: flex; 
+    align-items: center;
+  }
+  
+  .error-icon{
+    color: red;
+  }
+  
+  .message-error{
+    margin: auto 5px;
+  }
+  
+  .error-container{
+    margin-top: 5px;
+  }
+  
+  .close-icon{
+    color: #b4b4b4;
+  }
+  
+  
+  .close-icon:hover{
+    color: black;
+  }
